Moved driver-specific code to a separate file
